move player files to their own dir
