+++ src/amp-story-player/OWNERS
@@ -1,0 +1,17 @@
+// For an explanation of the OWNERS rules and syntax, see:
+// https://github.com/ampproject/amp-github-apps/blob/master/owners/OWNERS.example
+
+{
+  rules: [
+    {
+      owners: [
+        {
+          name: 'enriqe',
+          notify: true,
+        },
+        {name: 'gmajoulet'},
+        {name: 'newmuis'},
+      ],
+    },
+  ],
+}
