+++ src/amp-story-player/amp-story-player.js
