+++ test/unit/test-amp-story-player.js
@@ -14,8 +14,8 @@
  * limitations under the License.
  */
 
-import {AmpStoryPlayerManager} from '../../src/amp-story-player-manager';
-import {IFRAME_IDX} from '../../src/amp-story-player-impl';
+import {AmpStoryPlayerManager} from '../../src/amp-story-player/amp-story-player-manager';
+import {IFRAME_IDX} from '../../src/amp-story-player/amp-story-player-impl';
 import {Messaging} from '@ampproject/viewer-messaging';
 import {toArray} from '../../src/types';
